[SP-2] ♻️  Refactor. ProjectMainImage 확장성 개선, ImageGallery 위치 이동
- ImageGallery는 전역적을 사용 될 여지가 있기에 src/components 폴더로 이동
- contents/components 는 mdx에서 직접적으로 불러오는 컴포넌트들만 위치
