+++ contents/project/components/index.ts
@@ -1,0 +1,1 @@
+export { ProjectMainImage } from './main-image/index';
