--- contents/project/components/main-image/styles.ts
@@ -1,1 +1,0 @@
-export const mainGridStyle = `grid-cols-[1.6fr_repeat(2,_1fr)]`;
