+++ contents/project/components/main-image/types.ts
@@ -1,3 +1,3 @@
-import type { Caption } from '@contents/project/components/img-gallery/main';
+import type { Caption } from '@/components/img-gallery/main';
 
 export type MainImageItemContents = { caption?: Caption; src: string; className?: string };
