+++ src/components/img-gallery/index.ts
